src/main/java/Cohort.java
import java.util.ArrayList;
import java.util.List;

public class Cohort {

    private static List<Student> students;

    public Cohort() {
        this.students = new ArrayList<>();
    }

    public double getCohortAverage() {
        double avg = 0;
        for (Student student: this.getStudents()) {
            avg += student.getGradeAverage();
        }
        return avg / this.getStudents().size();
    }

    public void addStudent(Student student){
        students.add(student);
    }

    public static List<Student> getStudents() {
        return students;
    }
}